BUG: move tests to test_instrument
